pyproject.toml
[project]
name = "camilladsp-autoswitch"
version = "0.1.0"
description = "Automatic and safe profile switching for CamillaDSP"
requires-python = ">=3.9"

dependencies = [
  "PyYAML>=6.0"
]

[project.scripts]
cdspctl = "camilladsp_autoswitch.cli:main"
